README: Remove images and directly link to Discord CDN instead
There's no need to keep a copy of each yak emoji in this repo as well,
they're already in the SerenityOS/artwork repository. There's also no
note about their license anywhere in this repo, which is *not* the MIT
license as implied by the LICENSE file.
Just hotlink to the images on cdn.discordapp.com, which might also keep
them up-to-date automatically if changed (if the ID persists, that is).